+++ cookiecutter/utils.py
@@ -8,9 +8,13 @@ cookiecutter.utils
 Helper functions used throughout Cookiecutter.
 """
 
+from __future__ import unicode_literals
 import errno
+import logging
 import os
 import sys
+import contextlib
+
 
 PY3 = sys.version > '3'
 if PY3:
@@ -24,6 +28,8 @@ def make_sure_path_exists(path):
     Ensures that a directory exists.
     :param path: A directory path.
     """
+    
+    logging.debug("Making sure path exists: {0}".format(path))
     try:
         os.makedirs(path)
     except OSError as exception:
@@ -41,3 +47,18 @@ def unicode_open(filename, *args, **kwargs):
     if PY3:
         return open(filename, *args, **kwargs)
     return codecs.open(filename, *args, **kwargs)
+
+
+@contextlib.contextmanager
+def work_in(dirname=None):
+    """
+    Context manager version of os.chdir. When exited, returns to the working
+    directory prior to entering.
+    """
+    curdir = os.getcwd()
+    try:
+        if dirname is not None:
+            os.chdir(dirname)
+        yield
+    finally:
+        os.chdir(curdir)